Make the computation of the coordination sequence faster
The coordination sequence is now faster by precomputing all the
coordination spheres once for all the framework atoms in the beginng of
the analysis.
In addition a classmethod was added in atomic_parameters.py to compute
the maximum possible bond length that can be dected. This is used to exclude
any pair of atoms from any further consideration when they are separated by
a distance larger than the maximum possible bond lenght.