Create a MofStructure class.
Create a MofStructure class which extends the pymatgen (v 4.7.2) Structure class.
The Structure class was extended to include the self.metal and
self.organic properties. Methods from open_metal_detector.py
were coppied to enable the partion of a MOF into metal and organic part
and to compute the first coordination sphere for all metal atoms.