Create a MetalSite class, update MofStructure
Add a new class derived from pymatgen (v 4.7.2) Structure class called
MetalSite. This new class holds all the information for each metal site
and its coordination sphere as well as the methods to detect whether it
is open or not.
Update the MofStructure to include an analyse_metals method that checks
for all metals if they are open  using the newly added MetalSite class
and if they are unique using the newly added find_coordination_sequence
method.
All the new methods are copied over from the open_metal_detector.py file.