Change open metal site detection method.
The new procedure to detect if a metal site is open by examing the
sites in its coordination sphere is as following:

1) For all 3-atom combinations in the coordination sphere,
compute the plane that passes through them.
2) For each atom in the coordination sphere find which side of the
plane it falls (-1 or 1 and 0 if it falls on the plane).
3) A metal site is marked as open if:
    a) All other sites fall on the plane.
    b) The metal site falls on the plane and all other sites fall on
    one side of the plane.
    c) The metal site falls on one side of the plane and all other
    sites fall on the oposite side of the plane.

To dermine if a site falls on a plane, the angles defined by the
points (S, Pi , Sp) are computed, where S is the site in question,
Sp is the projection of the site on the plane and Pi is each of the
three sites that define the plane. If all three angles are less than
the tolerance value then the site is considered to be on the plane.
The default tolerance value is 15 degrees and can be changed by setting
the tolerance parameter when creating a MetalSite object.

To determine which side of a plane a site fals the sign of the
perpendicular distance from the plane is used.