Improve OMS analysis progress output,fix CIF check
Improve how the progress for the OMS analysis is presented.
Show the progress for all the batches in one line and
update that line. To do this a shared memory array is
used that tracks which mof is being analyzed by each
process. Then a loop updates the progress line based
on this array.

Also fix an issue that occured if a CIF file was not read
and was checked to see if it has no metal atoms.